+++ task_1_5_1/src/main/java/ru/nsu/fit/tretyakov/OperatorFactory.java
@@ -1,9 +1,9 @@
 package ru.nsu.fit.tretyakov;
 
-import ru.nsu.fit.tretyakov.operators.*;
-
 import java.util.ArrayList;
 import java.util.List;
+import ru.nsu.fit.tretyakov.operators.*;
+
 
 /**
  * This is the factory-class which produces specific operator by
