+++ task_1_5_1/src/test/java/ru/nsu/fit/tretyakov/test_operators/TernaryMultiply.java
@@ -1,16 +1,16 @@
-package ru.nsu.fit.tretyakov.testOperators;
+package ru.nsu.fit.tretyakov.test_operators;
+
+import java.util.Deque;
 
 import ru.nsu.fit.tretyakov.Number;
 import ru.nsu.fit.tretyakov.Operator;
 
-import java.util.Deque;
-
 /**
- * This class calculates adding value of three operands.
+ * This class calculates multiply value of three operands.
  */
-public class TernaryPlus implements Operator {
+public class TernaryMultiply implements Operator {
 
-    public static final String token = "+++";
+    public static final String token = "***";
 
     @Override
     public String getToken() {
@@ -18,16 +18,16 @@ public class TernaryPlus implements Operator {
     }
 
     /**
-     * This function calculates result of adding of three operands.
+     * This function calculates result of dividing three numbers.
      *
      * @param expressionStack is the stack of all operands in which
      *                        this function will get an operand.
-     * @return result of adding three numbers
+     * @return result of dividing three numbers.
      */
     @Override
     public Number calculate(Deque<Number> expressionStack) {
         Number[] operands = getValuesFromStack(3, expressionStack);
-        return new Number(operands[0].real() + operands[1].real()
-                + operands[2].real(), 0);
+        return new Number(operands[0].real() * operands[1].real()
+                * operands[2].real(), 0);
     }
 }
